Return repeat slot planning decisions
